refactor(cmd): extract HTTP assembly from main.go into internal/httpserver (#895)
Move the HTTP mux/route assembly, CORS, drain/shutdown sequencing, and the
adapter types out of cmd/mcp-data-platform/main.go (a second composition root)
into a new internal/httpserver package. main.go drops from 1,548 to 250 lines
and now holds only flag parsing, config load, platform construction, signal
handling, and the run/shutdown loop; it delegates to httpserver.Serve.

Package layout:
- server.go   Serve (was startHTTPServer; signature opts->address) + HTTP
              lifecycle (listen, drain, session-close, CORS, root/SSE handlers)
- mounts.go   mount*/build* handlers and adapter types (compositeOperation
              resolver, portalSearchAdapter, admin/gateway/observability/portal
              UI/browser-auth/OAuth-route wiring)
- wiring.go   connoauth refresher resolver, parseDurationWithDays, enrichment
              source wiring
- dbmounts.go mountPortalAPI/mountResourcesAPI, isolated because their bodies
              run only against a live Postgres

Move is behavior-preserving: the route/mount registration sequence is identical
(only opts.address -> address changed). HTTP tests move into internal/httpserver
alongside the code they test; main-level flag/run/migrate tests stay.

Supporting changes:
- .golangci.yml: admin-is-a-leaf now permits internal/httpserver (the extracted
  composition root) to import pkg/admin
- testdata/allowed_internal_imports.txt: regenerated for the new package edges
- CLAUDE.md: document internal/httpserver in the structure map
- pkg/platform/lifecycle.go: update a stale startHTTPServer doc reference
- internal/testdb: add NewWithDSN so an integration test can hand the DSN to
  platform.New

Coverage: dbmounts.go's two DB-backed mount functions cannot be exercised by the
unit suite (portal/resource stores need a real Postgres, confined to the
//go:build integration suite which emits no coverage profile). They are excluded
from scripts/patch-coverage.sh and codecov.yml with the same rationale as
cmd/dev-mcp-mock, and exercised by new RealDB integration tests
(dbmounts_realdb_integration_test.go). Patch coverage on the remaining moved
lines is 86.5%.

make verify passes (lint, patch-coverage, test-realdb, import-ratchet,
structure-map, GoReleaser dry-run).

Closes #895
