+++ cmd/mcp-data-platform/main.go
@@ -3,76 +3,33 @@ package main
 
 import (
 	"context"
-	"errors"
 	"flag"
 	"fmt"
 	"io"
-	"log"
 	"log/slog"
-	"net/http"
 	"os"
 	"os/signal"
-	"slices"
-	"strconv"
-	"strings"
 	"syscall"
 	"time"
 
-	sdkauth "github.com/modelcontextprotocol/go-sdk/auth"
 	"github.com/modelcontextprotocol/go-sdk/mcp"
-	"github.com/modelcontextprotocol/go-sdk/oauthex"
-	trinoclient "github.com/txn2/mcp-trino/pkg/client"
 
-	_ "github.com/txn2/mcp-data-platform/internal/apidocs" // Swagger API docs
+	"github.com/txn2/mcp-data-platform/internal/httpserver"
 	mcpserver "github.com/txn2/mcp-data-platform/internal/server"
-	"github.com/txn2/mcp-data-platform/internal/ui"
-	"github.com/txn2/mcp-data-platform/pkg/admin"
-	"github.com/txn2/mcp-data-platform/pkg/audit"
-	"github.com/txn2/mcp-data-platform/pkg/browsersession"
-	"github.com/txn2/mcp-data-platform/pkg/connoauth"
-	"github.com/txn2/mcp-data-platform/pkg/gatewayhttp"
-	"github.com/txn2/mcp-data-platform/pkg/health"
-	httpauth "github.com/txn2/mcp-data-platform/pkg/http"
-	"github.com/txn2/mcp-data-platform/pkg/knowledge"
-	"github.com/txn2/mcp-data-platform/pkg/observability/proxy"
-	"github.com/txn2/mcp-data-platform/pkg/persona"
-	"github.com/txn2/mcp-data-platform/pkg/pkcestore"
 	"github.com/txn2/mcp-data-platform/pkg/platform"
-	"github.com/txn2/mcp-data-platform/pkg/portal"
-	"github.com/txn2/mcp-data-platform/pkg/portal/datahubapi"
-	"github.com/txn2/mcp-data-platform/pkg/ratelimit"
-	"github.com/txn2/mcp-data-platform/pkg/registry"
-	"github.com/txn2/mcp-data-platform/pkg/resource"
-	"github.com/txn2/mcp-data-platform/pkg/session"
-	apigatewaykit "github.com/txn2/mcp-data-platform/pkg/toolkits/apigateway"
-	datahubkit "github.com/txn2/mcp-data-platform/pkg/toolkits/datahub"
-	gatewaykit "github.com/txn2/mcp-data-platform/pkg/toolkits/gateway"
-	"github.com/txn2/mcp-data-platform/pkg/toolkits/gateway/enrichment"
-	"github.com/txn2/mcp-data-platform/pkg/toolkits/gateway/sources"
-	trinokit "github.com/txn2/mcp-data-platform/pkg/toolkits/trino"
 )
 
 const (
-	defaultReadHeaderTimeout = 10 * time.Second
-	fallbackGracePeriod      = 25 * time.Second
-	fallbackPreShutdownDelay = 2 * time.Second
+	transportHTTP = "http"
 	// lifecycleStopTimeout bounds how long Platform.Stop is allowed to
 	// run before Close proceeds anyway. Sized so the full shutdown
 	// budget (preDelay + httpGrace + lifecycleStop + close overhead)
 	// fits inside a 60s terminationGracePeriodSeconds with headroom.
 	lifecycleStopTimeout = 10 * time.Second
-	transportHTTP        = "http"
 )
 
 // logKeyError is the structured-log key for an error value.
 const logKeyError = "error"
-
-// sessionDrainSettle is how long in-flight requests are given to finish before
-// live MCP sessions are closed on shutdown (#675). Long-lived SSE and streamable
-// streams never go idle, so until the sessions are closed the agent stays on the
-// old build; closing them drops the stream so Claude Code auto-reconnects to the
-// new build. A var so tests can lower it; capped at the grace period at use.
-var sessionDrainSettle = 3 * time.Second
 
 func main() {
 	if len(os.Args) > 1 && os.Args[1] == "migrate-config" {
@@ -237,365 +194,15 @@ func startServer(ctx context.Context, mcpServer *mcp.Server, p *platform.Platfor
 		return nil
 	case transportHTTP, "sse":
 		// HTTP serves both SSE (/sse, /message) and Streamable HTTP (/mcp).
-		// "sse" is accepted for backward compatibility.
-		return startHTTPServer(ctx, mcpServer, p, opts)
+		// "sse" is accepted for backward compatibility. The HTTP mux
+		// assembly and drain/shutdown sequencing live in internal/httpserver.
+		if err := httpserver.Serve(ctx, mcpServer, p, opts.address); err != nil {
+			return fmt.Errorf("running http server: %w", err)
+		}
+		return nil
 	default:
 		return fmt.Errorf("unknown transport: %s", opts.transport)
 	}
-}
-
-// corsMiddleware adds CORS headers for browser-based MCP clients.
-func corsMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		origin := r.Header.Get("Origin")
-		if origin == "" {
-			origin = "*"
-		}
-		w.Header().Set("Access-Control-Allow-Origin", origin)
-		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, DELETE, OPTIONS")
-		w.Header().Set("Access-Control-Allow-Headers",
-			"Content-Type, Authorization, Accept, X-API-Key, "+
-				"Mcp-Session-Id, Mcp-Protocol-Version, Last-Event-ID")
-		w.Header().Set("Access-Control-Expose-Headers", "Mcp-Session-Id")
-		w.Header().Set("Access-Control-Allow-Credentials", "true")
-		w.Header().Set("Access-Control-Max-Age", "86400")
-
-		if r.Method == "OPTIONS" {
-			w.WriteHeader(http.StatusOK)
-			return
-		}
-		next.ServeHTTP(w, r)
-	})
-}
-
-// httpConfig holds configuration extracted from the platform for HTTP servers.
-type httpConfig struct {
-	requireAuth   bool
-	portalUI      bool
-	tlsEnabled    bool
-	tlsCertFile   string
-	tlsKeyFile    string
-	streamableCfg platform.StreamableConfig
-	shutdownCfg   platform.ShutdownConfig
-	// mcpServer is closed-out on shutdown so connected agents reconnect to the new
-	// build (#675). Carried on the config so listenAndServe stays within its arg budget.
-	mcpServer *mcp.Server
-}
-
-func extractHTTPConfig(p *platform.Platform) httpConfig {
-	var cfg httpConfig
-	if p != nil && p.Config() != nil {
-		c := p.Config()
-		cfg.requireAuth = !c.Auth.AllowAnonymous
-		cfg.portalUI = (c.Portal.Enabled == nil || *c.Portal.Enabled) && ui.Available()
-		cfg.tlsEnabled = c.Server.TLS.Enabled
-		cfg.tlsCertFile = c.Server.TLS.CertFile
-		cfg.tlsKeyFile = c.Server.TLS.KeyFile
-		cfg.streamableCfg = c.Server.Streamable
-		cfg.shutdownCfg = c.Server.Shutdown
-	}
-	return cfg
-}
-
-// newSSEHandler creates an SSE handler with auth middleware.
-func newSSEHandler(mcpServer *mcp.Server, requireAuth bool, resourceMetadataURL string) http.Handler {
-	sseHandler := mcp.NewSSEHandler(func(*http.Request) *mcp.Server {
-		return mcpServer
-	}, nil)
-
-	if requireAuth && resourceMetadataURL != "" {
-		return httpauth.RequireAuthWithOAuth(resourceMetadataURL)(sseHandler)
-	}
-	if requireAuth {
-		return httpauth.RequireAuth()(sseHandler)
-	}
-	return httpauth.OptionalAuth()(sseHandler)
-}
-
-// resourceMetadataURL returns the protected resource metadata URL if OAuth is
-// enabled, or empty string otherwise.
-func resourceMetadataURL(p *platform.Platform) string {
-	if p == nil || p.OAuthServer() == nil {
-		return ""
-	}
-	return p.Config().OAuth.Issuer + "/.well-known/oauth-protected-resource"
-}
-
-func startHTTPServer(ctx context.Context, mcpServer *mcp.Server, p *platform.Platform, opts serverOptions) error {
-	// Gateway/api-gateway integrations (broadcaster, token stores, catalog
-	// store, embed-job queue) and the admin self-connection are wired by
-	// p.WireRuntime in startServer, which runs before this HTTP setup. That
-	// keeps their ordering in one place (#854): the SSE long-poll path still
-	// gets the broadcaster wired into the gateway toolkit before the listener
-	// comes up at the end of this function, so AwareHandler never accepts a
-	// subscriber ahead of the wiring.
-	if p != nil {
-		// Start the background OAuth refresher once toolkits and
-		// connection store are wired. Single-call here (not in the
-		// platform constructor) so the resolver can read the live
-		// ConnectionStore + OAuthKindHandlers from the fully-set-up
-		// platform — these are not available at platform.New time.
-		startConnOAuthRefresher(p)
-	}
-
-	mux := http.NewServeMux()
-	hcfg := extractHTTPConfig(p)
-	hc := health.NewChecker()
-
-	if !hcfg.tlsEnabled {
-		log.Println("WARNING: HTTP transport without TLS - credentials may be transmitted in plaintext")
-	}
-
-	// Health endpoints (registered before catch-all /)
-	mux.Handle("/healthz", hc.LivenessHandler())
-	mux.Handle("/readyz", hc.ReadinessHandler())
-
-	// Robots.txt — prevent search engines from indexing the portal.
-	mux.HandleFunc("GET /robots.txt", func(w http.ResponseWriter, _ *http.Request) {
-		w.Header().Set("Content-Type", "text/plain")
-		_, _ = fmt.Fprint(w, "User-agent: *\nDisallow: /\n")
-	})
-
-	// Mount OAuth server if enabled
-	if p != nil && p.OAuthServer() != nil {
-		registerOAuthRoutes(mux, p.OAuthServer())
-		log.Println("OAuth server enabled")
-	}
-
-	// Mount OAuth protected resource metadata (RFC 9728) when OAuth is
-	// enabled. MCP clients discover the authorization server from this
-	// endpoint after receiving an HTTP 401 with WWW-Authenticate header.
-	rmURL := resourceMetadataURL(p)
-	if rmURL != "" {
-		issuer := p.Config().OAuth.Issuer
-		mux.Handle("/.well-known/oauth-protected-resource",
-			sdkauth.ProtectedResourceMetadataHandler(&oauthex.ProtectedResourceMetadata{
-				Resource:               issuer,
-				AuthorizationServers:   []string{issuer},
-				BearerMethodsSupported: []string{"header"},
-				ResourceName:           p.Config().Server.Name,
-			}))
-		log.Println("OAuth protected resource metadata enabled on /.well-known/oauth-protected-resource")
-	}
-
-	// Mount browser auth routes (OIDC login/callback/logout)
-	mountBrowserAuth(mux, p)
-
-	// Mount admin API if enabled
-	mountAdminAPI(mux, p)
-
-	// The built-in platform-admin self-connection (issue #543) that lets an
-	// admin drive /api/v1/admin/* through the api gateway is seeded by
-	// p.WireRuntime (startServer), after the gateway integrations it depends
-	// on. The admin API mounted just above is the loopback surface it targets.
-
-	// Mount portal API if enabled
-	if err := mountPortalAPI(mux, p); err != nil {
-		return err
-	}
-
-	// Mount managed resources API if enabled
-	mountResourcesAPI(mux, p)
-
-	// Mount the REST gateway shim if an apigateway toolkit is loaded.
-	// Exposes api_invoke_endpoint over plain HTTP for non-MCP clients
-	// (e.g. Apache NiFi). Auth + persona + audit all flow through the
-	// MCP middleware chain via an in-memory session.
-	mountGatewayAPI(mux, mcpServer, p, hcfg.requireAuth)
-
-	// Mount the authenticated PromQL query proxy the portal's
-	// observability views read from (#462). Always mounted; returns 503
-	// when Prometheus is not configured.
-	mountObservabilityProxy(mux, p, hcfg.requireAuth)
-
-	// Mount unified portal UI (includes both portal and admin sections)
-	mountPortalUI(mux, p, ui.Available())
-
-	// Mount SSE handler (legacy clients)
-	wrappedSSE := newSSEHandler(mcpServer, hcfg.requireAuth, rmURL)
-	mux.Handle("/sse", wrappedSSE)
-	mux.Handle("/message", wrappedSSE)
-	log.Println("SSE transport enabled on /sse, /message")
-
-	// Build and mount the root handler (MCP streamable HTTP + session + browser redirect).
-	rootHandler := buildRootHandler(mcpServer, p, hcfg)
-	mountRootHandler(mux, rootHandler, hcfg, rmURL)
-
-	hcfg.mcpServer = mcpServer
-	return listenAndServe(ctx, opts.address, corsMiddleware(mux), hcfg, hc)
-}
-
-// buildRootHandler constructs the MCP streamable HTTP handler with optional
-// session-aware wrapping. Browser redirect is applied in mountRootHandler
-// so it wraps outside the auth gateway.
-func buildRootHandler(mcpServer *mcp.Server, p *platform.Platform, hcfg httpConfig) http.Handler {
-	streamableHandler := mcp.NewStreamableHTTPHandler(func(*http.Request) *mcp.Server {
-		return mcpServer
-	}, &mcp.StreamableHTTPOptions{
-		SessionTimeout: hcfg.streamableCfg.SessionTimeout,
-		Stateless:      hcfg.streamableCfg.Stateless,
-	})
-
-	// Wrap with AwareHandler when using external session store
-	// (database mode forces Stateless: true on the SDK, and sessions
-	// are managed by our handler against the external store).
-	var handler http.Handler = streamableHandler
-	if p != nil && p.SessionStore() != nil && hcfg.streamableCfg.Stateless {
-		handler = session.NewAwareHandler(streamableHandler, session.HandlerConfig{
-			Store:       p.SessionStore(),
-			TTL:         p.Config().Sessions.TTL,
-			Broadcaster: p.Broadcaster(),
-		})
-		// Platform.Broadcaster() is non-nil after New (the sessionsync
-		// layer wires postgres or memory). The "+ broadcaster" tag is part
-		// of the log line so operators can grep deployments where the
-		// session-aware handler is wired with the SSE long-poll path.
-		log.Println("Session-aware handler enabled (external session store + broadcaster)")
-	}
-
-	return handler
-}
-
-// mountRootHandler registers the root handler on the mux, optionally wrapping
-// it with the MCP auth gateway when authentication is required.
-// Browser redirect wraps OUTSIDE the auth gateway so that browser requests
-// (Accept: text/html) redirect to /portal/ without hitting the 401.
-func mountRootHandler(mux *http.ServeMux, rootHandler http.Handler, hcfg httpConfig, rmURL string) {
-	handler := rootHandler
-	if hcfg.requireAuth {
-		handler = httpauth.MCPAuthGateway(rmURL)(handler)
-		log.Println("Streamable HTTP transport enabled on / (auth required)")
-	} else {
-		log.Println("Streamable HTTP transport enabled on / (anonymous)")
-	}
-
-	if hcfg.portalUI {
-		handler = browserRedirectMiddleware(handler)
-	}
-
-	mux.Handle("/", handler)
-}
-
-func listenAndServe(ctx context.Context, addr string, handler http.Handler, hcfg httpConfig, hc *health.Checker) error {
-	server := &http.Server{
-		Addr:              addr,
-		Handler:           handler,
-		ReadHeaderTimeout: defaultReadHeaderTimeout,
-	}
-
-	gracePeriod := hcfg.shutdownCfg.GracePeriod
-	if gracePeriod == 0 {
-		gracePeriod = fallbackGracePeriod
-	}
-	preDelay := hcfg.shutdownCfg.PreShutdownDelay
-	if preDelay == 0 {
-		preDelay = fallbackPreShutdownDelay
-	}
-
-	go func() { // #nosec G118 -- ctx is the application-level shutdown context, not a request-scoped context
-		<-ctx.Done()
-
-		// Mark not-ready so K8s load balancer stops sending traffic.
-		if hc != nil {
-			hc.SetDraining()
-			slog.Info("shutdown: readiness set to draining, waiting for LB deregistration",
-				"pre_shutdown_delay", preDelay)
-			time.Sleep(preDelay)
-		}
-
-		slog.Info("shutdown: draining HTTP connections", "grace_period", gracePeriod)
-		drainHTTPServer(server, hcfg.mcpServer, gracePeriod)
-	}()
-
-	// Mark ready just before we start accepting connections.
-	if hc != nil {
-		hc.SetReady()
-	}
-
-	if hcfg.tlsEnabled {
-		log.Printf("Starting HTTP server with TLS on %s\n", addr)
-		if err := server.ListenAndServeTLS(hcfg.tlsCertFile, hcfg.tlsKeyFile); err != http.ErrServerClosed {
-			return fmt.Errorf("listening with TLS on %s: %w", addr, err)
-		}
-		return nil
-	}
-
-	log.Printf("Starting HTTP server on %s\n", addr)
-	if err := server.ListenAndServe(); err != http.ErrServerClosed {
-		return fmt.Errorf("listening on %s: %w", addr, err)
-	}
-	return nil
-}
-
-// drainHTTPServer gracefully drains the HTTP server, then closes lingering MCP
-// sessions so connected agents reconnect to the new build (#675). Long-lived
-// SSE/streamable streams never go idle, so server.Shutdown does not complete until
-// its grace deadline and the agent stays on the old build the whole time. It runs
-// concurrently: in-flight requests get a brief settle, then the sessions are closed
-// to drop those streams and trigger client auto-reconnect.
-func drainHTTPServer(server *http.Server, mcpServer *mcp.Server, gracePeriod time.Duration) {
-	shutdownCtx, cancel := context.WithTimeout(context.Background(), gracePeriod)
-	defer cancel()
-
-	done := make(chan error, 1)
-	go func() { done <- server.Shutdown(shutdownCtx) }()
-
-	settle := min(sessionDrainSettle, gracePeriod)
-	select {
-	case err := <-done:
-		logHTTPDrainResult(err)
-	case <-time.After(settle):
-		closeMCPSessions(shutdownCtx, mcpServer)
-		logHTTPDrainResult(<-done)
-	}
-}
-
-// closeMCPSessions closes every live MCP session so connected clients drop their
-// stale connection and reconnect to the new build (#675). Claude Code auto-reconnects
-// HTTP/SSE servers and re-handshakes (fresh tools/list); Claude Desktop requires an
-// app restart.
-//
-// ServerSession.Close is graceful: an idle session's long-lived SSE/streamable stream
-// drops immediately, but a session with an in-flight tool call blocks until that call
-// returns, and that wait is not bounded by the HTTP grace period. So the closes run in
-// a goroutine bounded by ctx (the shutdown deadline): if they do not all finish in
-// time, we return and let process exit drop the remaining connections rather than hang
-// past terminationGracePeriodSeconds and risk a SIGKILL mid-call.
-func closeMCPSessions(ctx context.Context, mcpServer *mcp.Server) {
-	if mcpServer == nil {
-		return
-	}
-	var sessions []*mcp.ServerSession
-	for s := range mcpServer.Sessions() {
-		sessions = append(sessions, s)
-	}
-	if len(sessions) == 0 {
-		return
-	}
-
-	closed := make(chan struct{})
-	go func() {
-		for _, s := range sessions {
-			_ = s.Close() // returns once the session's in-flight requests finish
-		}
-		close(closed)
-	}()
-
-	select {
-	case <-closed:
-		slog.Info("shutdown: closed live MCP sessions so clients reconnect to the new build", "count", len(sessions))
-	case <-ctx.Done():
-		slog.Warn("shutdown: MCP session close did not finish before the grace deadline; process exit will drop remaining connections", "count", len(sessions))
-	}
-}
-
-// logHTTPDrainResult records the outcome of the HTTP server drain.
-func logHTTPDrainResult(err error) {
-	if err != nil {
-		slog.Error("shutdown: HTTP drain error", logKeyError, err)
-		return
-	}
-	slog.Info("shutdown: HTTP server stopped")
 }
 
 // stdioMarker is the conventional marker for stdin/stdout in CLI tools.
@@ -640,909 +247,4 @@ func runMigrateConfig(args []string) error {
 		return fmt.Errorf("migrating config: %w", err)
 	}
 	return nil
-}
-
-// defaultAdminPathPrefix and defaultAdminPersona mirror applyAdminDefaults for
-// configs injected directly (via WithConfig) that never ran through
-// applyDefaults. Now that admin defaults to enabled, an empty prefix would mount
-// the admin API at "/" and collide with the root MCP handler, and an empty
-// persona would reject every admin request (buildAdminHandler compares the
-// caller's persona against it), locking admins out.
-const (
-	defaultAdminPathPrefix = "/api/v1/admin"
-	defaultAdminPersona    = "admin"
-)
-
-// mountAdminAPI registers the admin REST API on the mux if enabled.
-func mountAdminAPI(mux *http.ServeMux, p *platform.Platform) {
-	if p == nil || !p.Config().Admin.IsEnabled() {
-		return
-	}
-	adminHandler := buildAdminHandler(p)
-	prefix := p.Config().Admin.PathPrefix
-	if prefix == "" {
-		prefix = defaultAdminPathPrefix
-	}
-	mux.Handle(prefix+"/", adminHandler)
-	log.Println("Admin API enabled on", prefix)
-}
-
-// mountPortalAPI registers the portal REST API on the mux if portal is enabled.
-// portalDisabled returns true when portal is explicitly disabled or platform is nil.
-func portalDisabled(p *platform.Platform) bool {
-	if p == nil {
-		return true
-	}
-	e := p.Config().Portal.Enabled
-	return e != nil && !*e
-}
-
-func mountPortalAPI(mux *http.ServeMux, p *platform.Platform) error {
-	if p == nil || portalDisabled(p) {
-		return nil
-	}
-	if p.PortalAssetStore() == nil || p.PortalShareStore() == nil {
-		log.Println("Portal enabled but stores not available (database required)")
-		return nil
-	}
-
-	// Build the trusted-proxy-aware client-IP resolver for the public viewer's
-	// rate limiter (#904). A malformed CIDR is a boot-time configuration error
-	// surfaced here, the genuine startup path for the portal.
-	rlResolver, err := portalRateLimitResolver(p.Config().Portal.RateLimit)
-	if err != nil {
-		return err
-	}
-
-	var portalAuthOpts []portal.AuthenticatorOption
-	if p.BrowserSessionAuth() != nil {
-		portalAuthOpts = append(portalAuthOpts, portal.WithBrowserAuth(p.BrowserSessionAuth()))
-	}
-	portalAuth := portal.NewAuthenticator(p.Authenticator(), portalAuthOpts...)
-
-	var adminRoles []string
-	if pr := p.PersonaRegistry(); pr != nil {
-		if adminP, ok := pr.Get(p.Config().Admin.Persona); ok {
-			adminRoles = adminP.Roles
-		}
-	}
-
-	brandName := portalBrandName(p)
-
-	deps := portal.Deps{
-		AssetStore:                  p.PortalAssetStore(),
-		ShareStore:                  p.PortalShareStore(),
-		VersionStore:                p.PortalVersionStore(),
-		CollectionStore:             p.PortalCollectionStore(),
-		ThreadStore:                 p.PortalThreadStore(),
-		KnowledgePageStore:          p.PortalKnowledgePageStore(),
-		KnowledgePageDedupThreshold: p.Config().Knowledge.Pages.Resolve().DedupThreshold,
-		S3Client:                    p.PortalS3Client(),
-		S3Bucket:                    p.Config().Portal.S3Bucket,
-		PublicBaseURL:               p.Config().Portal.PublicBaseURL,
-		RateLimit: portal.RateLimitConfig{
-			RequestsPerMinute: p.Config().Portal.RateLimit.RequestsPerMinute,
-			BurstSize:         p.Config().Portal.RateLimit.BurstSize,
-		},
-		RateLimitResolver:  rlResolver,
-		OIDCEnabled:        p.BrowserSessionFlow() != nil,
-		AdminRoles:         adminRoles,
-		Authenticator:      portalAuth,
-		PromptStore:        p.PromptStore(),
-		PromptRegistrar:    p,
-		PromptInfoProvider: p,
-		BrandName:          brandName,
-		BrandLogoSVG:       p.BrandLogoSVG(),
-		BrandURL:           p.BrandURL(),
-		ImplementorName:    p.Config().Portal.Implementor.Name,
-		ImplementorLogoSVG: p.ResolveImplementorLogo(),
-		ImplementorURL:     p.Config().Portal.Implementor.URL,
-	}
-
-	wirePortalOptionalDeps(&deps, p)
-
-	handler := portal.NewHandler(deps, portal.RequirePortalAuth(portalAuth))
-	mux.Handle("/api/v1/portal/", handler)
-	mux.Handle("/portal/view/", handler)
-	log.Println("Portal API enabled on /api/v1/portal/")
-	return nil
-}
-
-// portalBrandName resolves the platform brand shown in the public viewer
-// header (far right): prefer the mcpapps platform-info config, then the portal
-// title, then the server name.
-func portalBrandName(p *platform.Platform) string {
-	if name := mcpappsBrandName(p); name != "" {
-		return name
-	}
-	if title := p.Config().Portal.Title; title != "" {
-		return title
-	}
-	return p.Config().Server.Name
-}
-
-// portalRateLimitResolver builds the public viewer's trusted-proxy-aware
-// client-IP resolver from the portal rate-limit config (#904). An empty
-// trusted-proxy list yields the safe trust-none default (direct peer address;
-// X-Forwarded-For ignored). A malformed CIDR is a configuration error wrapped
-// for the boot path. It also emits the untrusted-proxy footgun warning.
-func portalRateLimitResolver(cfg platform.PortalRateLimitConfig) (*ratelimit.Resolver, error) {
-	resolver, err := ratelimit.NewResolver(cfg.TrustedProxies)
-	if err != nil {
-		return nil, fmt.Errorf("portal rate limiter: %w", err)
-	}
-	warnOnUntrustedPortalRateLimit(cfg.TrustedProxies)
-	return resolver, nil
-}
-
-// warnOnUntrustedPortalRateLimit surfaces the portal rate-limiting footgun at
-// boot, mirroring the OAuth endpoints: with no trusted proxies configured,
-// client attribution falls back to the direct peer address. Behind a reverse
-// proxy or k8s ingress every client shares the proxy's IP, so the per-client
-// limit collapses onto one bucket. The global backstop still bounds total
-// load; the operator should set portal.rate_limit.trusted_proxies to restore
-// per-client fairness.
-func warnOnUntrustedPortalRateLimit(trustedProxies []string) {
-	if len(trustedProxies) == 0 {
-		log.Println("Portal rate limiting is on but portal.rate_limit.trusted_proxies is empty: " +
-			"behind a reverse proxy or ingress every client shares the proxy IP, so per-client " +
-			"limiting collapses to a single bucket. Set portal.rate_limit.trusted_proxies to your " +
-			"proxy/ingress CIDRs (the global backstop still bounds total load meanwhile).")
-	}
-}
-
-// mountResourcesAPI registers the managed resources REST API on the mux if enabled.
-func mountResourcesAPI(mux *http.ServeMux, p *platform.Platform) {
-	if p == nil || p.ResourceStore() == nil {
-		return
-	}
-
-	var portalAuthOpts []portal.AuthenticatorOption
-	if p.BrowserSessionAuth() != nil {
-		portalAuthOpts = append(portalAuthOpts, portal.WithBrowserAuth(p.BrowserSessionAuth()))
-	}
-	portalAuth := portal.NewAuthenticator(p.Authenticator(), portalAuthOpts...)
-
-	pr := p.PersonaRegistry()
-	adminPersona := p.Config().Admin.Persona
-	extractClaims := func(r *http.Request) (*resource.Claims, error) {
-		user, err := portalAuth.Authenticate(r)
-		if err != nil {
-			// Surface a CSRF rejection as resource.ErrForbidden so the handler
-			// responds 403 (recoverable) instead of 401 (force-logout), matching
-			// the admin/portal surfaces.
-			if errors.Is(err, browsersession.ErrCSRFInvalid) {
-				return nil, resource.ErrForbidden
-			}
-			return nil, errors.New("authentication required")
-		}
-		if user == nil {
-			return nil, errors.New("authentication required")
-		}
-		return buildResourceClaims(user, pr, adminPersona), nil
-	}
-
-	deps := resource.Deps{
-		Store:     p.ResourceStore(),
-		S3Client:  p.ResourceS3Client(),
-		S3Bucket:  p.Config().Resources.Managed.S3Bucket,
-		URIScheme: p.Config().Resources.Managed.URIScheme,
-		OnCreate:  p.RegisterManagedResource,
-		OnDelete:  p.UnregisterManagedResource,
-	}
-
-	handler := resource.NewHandler(deps, extractClaims, nil)
-	mux.Handle("/api/v1/resources/", handler)
-	mux.Handle("/api/v1/resources", handler)
-	log.Println("Managed resources API enabled on /api/v1/resources")
-}
-
-// mountGatewayAPI registers the REST shim for the apigateway toolkit
-// on the mux. The shim is only mounted when at least one apigateway
-// toolkit instance is loaded; otherwise the route would always return
-// "connection not found" and add noise to the route table.
-//
-// Auth wrapping mirrors the MCP root handler: when requireAuth is on,
-// the handler is wrapped with httpauth.RequireAuth so that requests
-// without a credential are rejected at the HTTP layer before reaching
-// the in-memory MCP session. When auth is off, the wrapper is a no-op
-// and the request flows through anonymously (matching the rest of the
-// platform's behavior in that mode).
-// compositeOperationResolver fans an operationId lookup across every
-// registered apigateway toolkit (a multi-instance deployment splits
-// connections across toolkits). The first non-empty result wins; an
-// empty result means no toolkit owns the connection or no spec path
-// matched, which the metrics middleware maps to "unknown".
-type compositeOperationResolver []*apigatewaykit.Toolkit
-
-func (c compositeOperationResolver) ResolveOperationID(ctx context.Context, connection, method, path string) string {
-	for _, tk := range c {
-		if op := tk.ResolveOperationID(ctx, connection, method, path); op != "" {
-			return op
-		}
-	}
-	return ""
-}
-
-func (c compositeOperationResolver) HasConnection(connection string) bool {
-	for _, tk := range c {
-		if tk.HasConnection(connection) {
-			return true
-		}
-	}
-	return false
-}
-
-func mountGatewayAPI(mux *http.ServeMux, mcpServer *mcp.Server, p *platform.Platform, requireAuth bool) {
-	if mcpServer == nil || p == nil {
-		return
-	}
-	apiToolkits := p.ToolkitRegistry().GetByKind(apigatewaykit.Kind)
-	if len(apiToolkits) == 0 {
-		return
-	}
-
-	resolver := make(compositeOperationResolver, 0, len(apiToolkits))
-	for _, tk := range apiToolkits {
-		if api, ok := tk.(*apigatewaykit.Toolkit); ok {
-			resolver = append(resolver, api)
-		}
-	}
-
-	handler, err := gatewayhttp.NewHandler(gatewayhttp.Deps{
-		MCPServer:   mcpServer,
-		Metrics:     p.Metrics(),
-		Resolver:    resolver,
-		Identity:    p.NewGatewayIdentityResolver(),
-		RawMaxBytes: p.APIGatewayRawMaxBytes(),
-	})
-	if err != nil {
-		log.Printf("REST gateway disabled: %v", err)
-		return
-	}
-
-	wrapped := handler
-	if requireAuth {
-		wrapped = httpauth.RequireAuth()(handler)
-	}
-	mux.Handle("/api/v1/gateway/", wrapped)
-	log.Println("REST gateway enabled on /api/v1/gateway/{connection}/invoke")
-}
-
-// defaultPrometheusURL is the auto-discovered in-cluster Prometheus endpoint
-// used when observability.prometheus.url is not configured. Set that config
-// value only to point at a Prometheus deployed under a different name.
-const defaultPrometheusURL = "http://mcp-data-platform-prometheus:9090"
-
-// mountObservabilityProxy mounts the authenticated PromQL query proxy at
-// /api/v1/observability/. It is always mounted (gated behind auth +
-// the observability:read persona capability); when Prometheus is not
-// configured its endpoints return 503 so the portal renders a clean
-// empty state.
-func mountObservabilityProxy(mux *http.ServeMux, p *platform.Platform, requireAuth bool) {
-	if p == nil {
-		return
-	}
-	pc := p.Config().Observability.Prometheus
-	if pc.URL == "" {
-		// Auto-discover the default in-cluster Prometheus; the config only
-		// needs to override this to point at a non-default deployment.
-		pc.URL = defaultPrometheusURL
-	}
-	handler, err := proxy.New(proxy.Config{
-		URL:                pc.URL,
-		Timeout:            pc.Timeout,
-		BasicAuthUser:      pc.BasicAuth.Username,
-		BasicAuthPass:      pc.BasicAuth.Password,
-		RateLimitPerSecond: pc.RateLimitPerSecond,
-	}, p.NewObservabilityAuthorizer())
-	if err != nil {
-		log.Printf("observability proxy disabled: %v", err)
-		return
-	}
-
-	proxyMux := http.NewServeMux()
-	handler.Register(proxyMux)
-
-	var wrapped http.Handler = proxyMux
-	if requireAuth {
-		// The portal SPA calls these endpoints directly with its
-		// browser-session cookie, so accept that (like the admin and
-		// portal APIs) in addition to Bearer/API-key tokens. The proxy's
-		// authorizer enforces authentication (401) and the
-		// observability:read capability (403); OptionalAuth only lifts a
-		// present token onto the context without rejecting cookie auth.
-		wrapped = p.ObservabilityAuthMiddleware()(httpauth.OptionalAuth()(proxyMux))
-	}
-	mux.Handle("/api/v1/observability/", wrapped)
-	if pc.URL == "" {
-		log.Println("observability proxy mounted (Prometheus not configured; endpoints return 503)")
-	} else {
-		log.Println("observability proxy enabled on /api/v1/observability/{query,query_range}")
-	}
-}
-
-// buildResourceClaims creates resource Claims from an authenticated user,
-// resolving persona memberships and admin status from the persona registry.
-func buildResourceClaims(user *portal.User, pr *persona.Registry, adminPersona string) *resource.Claims {
-	claims := &resource.Claims{
-		Sub:   user.UserID,
-		Email: user.Email,
-		Roles: user.Roles,
-	}
-	if pr != nil {
-		for _, per := range pr.All() {
-			if matchesAnyRole(per.Roles, user.Roles) {
-				claims.Personas = append(claims.Personas, per.Name)
-				if per.Name == adminPersona {
-					claims.IsAdmin = true
-				}
-			}
-		}
-	}
-	claims.AdminOfPersonas = extractPersonaAdminRoles(user.Roles)
-	return claims
-}
-
-// personaAdminInfix is the role substring that marks a persona-admin grant.
-// Roles may carry an arbitrary prefix (e.g., "dp_persona-admin:finance").
-const personaAdminInfix = "persona-admin:"
-
-// extractPersonaAdminRoles extracts persona names from roles containing
-// the "persona-admin:" pattern, tolerating any prefix.
-func extractPersonaAdminRoles(roles []string) []string {
-	var out []string
-	for _, r := range roles {
-		if _, name, ok := strings.Cut(r, personaAdminInfix); ok && name != "" {
-			out = append(out, name)
-		}
-	}
-	return out
-}
-
-// matchesAnyRole checks if any persona role matches any user role.
-func matchesAnyRole(personaRoles, userRoles []string) bool {
-	for _, pr := range personaRoles {
-		if slices.Contains(userRoles, pr) {
-			return true
-		}
-	}
-	return false
-}
-
-// mcpappsBrandName extracts brand_name from the mcpapps platform-info config,
-// or returns empty string if not configured.
-func mcpappsBrandName(p *platform.Platform) string {
-	appCfg, ok := p.Config().MCPApps.Apps["platform-info"]
-	if !ok {
-		return ""
-	}
-	name, _ := appCfg.Config["brand_name"].(string)
-	return name
-}
-
-// wirePortalOptionalDeps populates optional portal dependencies (audit, knowledge, persona).
-func wirePortalOptionalDeps(deps *portal.Deps, p *platform.Platform) {
-	if p.AuditStore() != nil {
-		deps.AuditMetrics = p.AuditStore()
-	}
-	if p.KnowledgeInsightStore() != nil {
-		deps.InsightStore = p.KnowledgeInsightStore()
-	}
-	if p.KnowledgeChangesetStore() != nil {
-		deps.ChangesetReader = p.KnowledgeChangesetStore()
-	}
-	if p.MemoryStore() != nil {
-		deps.MemoryStore = p.MemoryStore()
-		deps.MemoryWriter = p.MemoryStore()
-	}
-	if ep := p.EmbeddingProvider(); ep != nil {
-		deps.EmbeddingProvider = ep
-	}
-	if us := p.UserStore(); us != nil {
-		deps.UserDirectory = us
-	}
-	if pr := p.PersonaRegistry(); pr != nil {
-		tr := p.ToolkitRegistry()
-		deps.PersonaResolver = buildPersonaResolver(pr, tr)
-	}
-	if router := p.KnowledgeRouter(); router != nil {
-		deps.SearchRouter = portalSearchAdapter{router: router}
-	}
-	if reg := buildDataHubRegistrar(p, deps.PersonaResolver, deps.AdminRoles); reg != nil {
-		deps.DataHubRegistrar = reg
-	}
-}
-
-// buildDataHubRegistrar assembles the portal DataHub REST handler (#718) over the
-// live DataHub toolkit clients and returns its route registrar, or nil when no
-// DataHub connection is registered. The bridge reuses each toolkit's client for
-// both the semantic read adapter and the batched write client; a read-only
-// connection (read_only=true) exposes no writer.
-func buildDataHubRegistrar(p *platform.Platform, resolver portal.PersonaResolver, adminRoles []string) func(*http.ServeMux) {
-	if p.ToolkitRegistry() == nil {
-		return nil
-	}
-	urn := p.Config().Semantic.URNMapping
-	platformName := urn.Platform
-	if platformName == "" {
-		platformName = "trino"
-	}
-
-	bridge := datahubapi.NewStaticBridge()
-	for _, tk := range p.ToolkitRegistry().All() {
-		dhTk, ok := tk.(*datahubkit.Toolkit)
-		if !ok || dhTk.Client() == nil {
-			continue
-		}
-		reader, writer, err := datahubapi.BuildConnection(dhTk.Client(), platformName, urn.CatalogMapping, dhTk.Config().ReadOnly)
-		if err != nil {
-			log.Printf("portal datahub: skipping connection %q: %v", dhTk.Name(), err)
-			continue
-		}
-		bridge.Add(dhTk.Name(), reader, writer)
-	}
-	if bridge.Empty() {
-		return nil
-	}
-
-	var auditLogger audit.Logger
-	if store := p.AuditStore(); store != nil {
-		auditLogger = store
-	}
-	handler := datahubapi.NewHandler(datahubapi.Deps{
-		Bridge:          bridge,
-		PersonaResolver: resolver,
-		AdminRoles:      adminRoles,
-		Audit:           auditLogger,
-	})
-	return handler.Register
-}
-
-// portalSearchAdapter bridges the portal's GET /search endpoint to the unified
-// knowledge router. It lives in main (not pkg/portal) because pkg/knowledge
-// already imports pkg/portal for the asset and knowledge-page stores; the
-// adapter maps the portal-local search types to and from the knowledge types so
-// neither package needs to import the other.
-type portalSearchAdapter struct {
-	router *knowledge.Router
-}
-
-// Search forwards a portal search to the knowledge router, mapping the request
-// and response across the two type sets. No ranking or scope logic is added: the
-// router owns identity scoping, fusion, and allocation.
-func (a portalSearchAdapter) Search(ctx context.Context, q portal.SearchQuery) (portal.SearchResult, error) {
-	res, err := a.router.Search(ctx, knowledge.Query{
-		Intent:     q.Intent,
-		EntityURNs: q.EntityURNs,
-		Status:     q.Status,
-		Sources:    q.Sources,
-		Caller:     knowledge.Caller{UserID: q.Caller.UserID, Email: q.Caller.Email, Persona: q.Caller.Persona},
-		Limit:      q.Limit,
-	})
-	if err != nil {
-		return portal.SearchResult{}, fmt.Errorf("knowledge search: %w", err)
-	}
-	out := portal.SearchResult{Ranking: res.Ranking, UnknownSources: res.UnknownSources}
-	for _, g := range res.Groups {
-		hits := make([]portal.SearchHit, 0, len(g.Hits))
-		for _, hit := range g.Hits {
-			hits = append(hits, portal.SearchHit{
-				Text:       hit.Text,
-				Source:     hit.Source,
-				Ref:        hit.Ref,
-				Score:      hit.Score,
-				Status:     hit.Status,
-				EntityURNs: hit.EntityURNs,
-				Dimension:  hit.Dimension,
-			})
-		}
-		out.Groups = append(out.Groups, portal.SearchGroup{Source: g.Source, Hits: hits})
-	}
-	for _, c := range res.Coverage {
-		out.Coverage = append(out.Coverage, portal.SearchCoverage{Source: c.Source, Matched: c.Matched, Shown: c.Shown})
-	}
-	return out, nil
-}
-
-// buildPersonaResolver creates a portal.PersonaResolver from the persona and toolkit registries.
-func buildPersonaResolver(pr *persona.Registry, tr *registry.Registry) portal.PersonaResolver {
-	return func(roles []string) *portal.PersonaInfo {
-		per, ok := pr.GetForRoles(roles)
-		if !ok || per == nil {
-			return nil
-		}
-		info := &portal.PersonaInfo{Name: per.Name}
-		if tr != nil {
-			filter := persona.NewToolFilter(pr)
-			info.Tools = filter.FilterTools(per, tr.AllTools())
-		}
-		return info
-	}
-}
-
-// mountPortalUI registers the unified portal SPA frontend on the mux when the
-// portal UI config gate is enabled and assets are available.
-func mountPortalUI(mux *http.ServeMux, p *platform.Platform, assetsAvailable bool) {
-	if p == nil || portalDisabled(p) || !assetsAvailable {
-		return
-	}
-	mux.Handle("/portal/", http.StripPrefix("/portal", ui.Handler()))
-	log.Println("Portal UI enabled on /portal/")
-}
-
-// buildAdminHandler constructs the admin REST API handler from the platform.
-func buildAdminHandler(p *platform.Platform) http.Handler {
-	var authOpts []admin.PlatformAuthOption
-	if p.BrowserSessionAuth() != nil {
-		authOpts = append(authOpts, admin.WithBrowserSessionAuth(p.BrowserSessionAuth()))
-	}
-	adminPersona := p.Config().Admin.Persona
-	if adminPersona == "" {
-		adminPersona = defaultAdminPersona
-	}
-	platAuth := admin.NewPlatformAuthenticator(
-		p.Authenticator(),
-		adminPersona,
-		p.PersonaRegistry(),
-		authOpts...,
-	)
-
-	deps := admin.Deps{
-		Config:             p.Config(),
-		ConfigStore:        p.ConfigStore(),
-		FileDefaults:       p.FileDefaults(),
-		PersonaRegistry:    p.PersonaRegistry(),
-		ToolkitRegistry:    p.ToolkitRegistry(),
-		ReloadNotifier:     p,
-		MCPServer:          p.MCPServer(),
-		BrowserAuth:        p.BrowserSessionAuth(),
-		DatabaseAvailable:  p.Config().Database.DSN != "",
-		PlatformTools:      p.PlatformTools(),
-		AssetStore:         p.PortalAssetStore(),
-		ShareStore:         p.PortalShareStore(),
-		VersionStore:       p.PortalVersionStore(),
-		S3Client:           p.PortalS3Client(),
-		S3Bucket:           p.Config().Portal.S3Bucket,
-		ConnectionStore:    p.ConnectionStore(),
-		ConnectionSources:  p.ConnectionSources(),
-		EnrichmentStore:    p.EnrichmentStore(),
-		ToolkitsConfig:     p.Config().Toolkits,
-		PersonaStore:       p.PersonaStore(),
-		APIKeyStore:        p.APIKeyStore(),
-		UserStore:          p.UserStore(),
-		PromptStore:        p.PromptStore(),
-		PromptRegistrar:    p,
-		PromptInfoProvider: p,
-		FilePersonaNames:   p.FilePersonaNames(),
-	}
-
-	if p.AuditStore() != nil {
-		deps.AuditQuerier = p.AuditStore()
-		deps.AuditMetricsQuerier = p.AuditStore()
-	}
-
-	// Note: WireGatewayTokenStore and WireGatewayBroadcaster run earlier
-	// in startHTTPServer so they apply even when admin is disabled.
-	if engine := wireEnrichmentEngine(p); engine != nil {
-		deps.EnrichmentEngine = engine
-	}
-
-	// PKCE state: DB-backed (multi-replica safe) when a database is
-	// configured. otherwise an in-memory store with a background GC
-	// goroutine is wired here — single-replica only.
-	if db := p.DB(); db != nil {
-		deps.PKCEStore = pkcestore.NewPostgresStore(db, p.RestEncryptor())
-	} else {
-		deps.PKCEStore = pkcestore.NewMemoryStore()
-	}
-
-	// Wire the unified OAuth flow: shared connoauth.Store plus one
-	// OAuthKindHandler per connection kind. When both are present the
-	// admin handler activates the unified /connections/{kind}/{name}
-	// routes; otherwise it falls back to the legacy per-kind routes
-	// for backward compatibility during rollout.
-	deps.ConnOAuthStore = p.ConnOAuthStore()
-	deps.OAuthKinds = buildOAuthKindHandlers(p)
-	deps.AuthEvents = p.AuthEventWriter()
-	deps.AuthEventStore = p.AuthEventStore()
-	if catStore := p.APIGatewayCatalogStore(); catStore != nil {
-		deps.APICatalogStore = catStore
-	}
-	deps.Embedder = p.EmbeddingProvider()
-	if jobs := p.APIGatewayEmbedJobsStore(); jobs != nil {
-		deps.EmbedJobs = jobs
-	}
-	if reporter := p.IndexJobsReporter(); reporter != nil {
-		deps.IndexJobs = reporter
-	}
-
-	if p.KnowledgeInsightStore() != nil {
-		deps.Knowledge = admin.NewKnowledgeHandler(
-			p.KnowledgeInsightStore(),
-			p.KnowledgeChangesetStore(),
-			p.KnowledgeDataHubWriter(),
-			p.PortalKnowledgePageStore(),
-		)
-	}
-
-	if p.APIKeyAuthenticator() != nil {
-		deps.APIKeyManager = p.APIKeyAuthenticator()
-	}
-
-	return admin.NewHandler(deps, admin.RequirePersona(platAuth))
-}
-
-// connOAuthConfigResolver bridges the connoauth refresher's
-// ConfigResolver interface to the platform's ConnectionStore +
-// OAuthKindHandlers wiring. The refresher cannot import the platform
-// package directly (import cycle), so this adapter lives in main.go
-// where both packages are already imported.
-type connOAuthConfigResolver struct {
-	store     admin.ConnectionStore
-	kinds     admin.OAuthKindHandlers
-	maxLifeFn func(kind, name string, cfg map[string]any) time.Duration
-}
-
-// ResolveConfig fetches the connection_instances row for (kind, name)
-// and parses out the connoauth.Config via the per-kind handler. The
-// ErrConfigNotResolvable sentinel is returned when the connection no
-// longer exists OR is configured for a non-OAuth auth mode — the
-// refresher treats either as "skip" rather than "fail" so a stale
-// token row doesn't stall keepalive for other connections.
-func (r *connOAuthConfigResolver) ResolveConfig(ctx context.Context, key connoauth.Key) (connoauth.Config, error) {
-	handler, ok := r.kinds[key.Kind]
-	if !ok {
-		return connoauth.Config{}, connoauth.ErrConfigNotResolvable
-	}
-	inst, err := r.store.Get(ctx, key.Kind, key.Name)
-	if err != nil {
-		return connoauth.Config{}, connoauth.ErrConfigNotResolvable
-	}
-	cfg, err := handler.ParseOAuthConfig(inst.Config)
-	if err != nil {
-		return connoauth.Config{}, connoauth.ErrConfigNotResolvable
-	}
-	return cfg, nil
-}
-
-// MaxLifetime reads the per-connection oauth2_refresh_max_lifetime
-// config field. Zero when unset; the refresher then relies on
-// IdP-disclosed deadlines only (which is correct for Keycloak / Auth0
-// / Okta but inadequate for Microsoft / Salesforce / Google APIs
-// that don't disclose refresh-token deadlines but enforce wall-clock
-// max lifetimes anyway).
-func (r *connOAuthConfigResolver) MaxLifetime(ctx context.Context, key connoauth.Key) time.Duration {
-	if r.maxLifeFn == nil {
-		return 0
-	}
-	inst, err := r.store.Get(ctx, key.Kind, key.Name)
-	if err != nil {
-		return 0
-	}
-	return r.maxLifeFn(key.Kind, key.Name, inst.Config)
-}
-
-// readMaxLifetime extracts the operator-configured wall-clock max
-// lifetime for the refresh token, parsing the standard Go duration
-// string format ("60d" via the d-suffix helper). Returns zero when
-// the field is absent, empty, or unparseable — the refresher
-// gracefully degrades to IdP-disclosed-deadline-only mode in that
-// case.
-func readMaxLifetime(_, _ string, cfg map[string]any) time.Duration {
-	raw, _ := cfg[configKeyOAuthRefreshMaxLifetime].(string)
-	if raw == "" {
-		return 0
-	}
-	d, err := parseDurationWithDays(raw)
-	if err != nil {
-		return 0
-	}
-	return d
-}
-
-// configKeyOAuthRefreshMaxLifetime is the connection_instances
-// config key that holds the operator's wall-clock refresh-token max
-// lifetime hint. Stored as a duration string ("60d", "90d", "30d").
-const configKeyOAuthRefreshMaxLifetime = "oauth2_refresh_max_lifetime"
-
-// hoursPerDay names the magic number 24 so the lint rule on
-// numeric literals doesn't fire and so the math reads as intent.
-const hoursPerDay = 24
-
-// parseDurationWithDays is time.ParseDuration with a "d" suffix
-// shorthand added. The stdlib's time.ParseDuration tops out at "h",
-// so "60d" is unparseable. Refresh-token deadlines are routinely
-// expressed in days by operators (Microsoft 90d, Salesforce 30d), so
-// asking them to write "1440h" instead would be a thousand-cuts UX
-// failure.
-func parseDurationWithDays(s string) (time.Duration, error) {
-	if head, ok := strings.CutSuffix(s, "d"); ok {
-		days, err := strconv.Atoi(head)
-		if err != nil {
-			return 0, fmt.Errorf("parse duration %q: %w", s, err)
-		}
-		if days < 0 {
-			return 0, fmt.Errorf("parse duration %q: negative days", s)
-		}
-		return time.Duration(days) * hoursPerDay * time.Hour, nil
-	}
-	d, err := time.ParseDuration(s)
-	if err != nil {
-		return 0, fmt.Errorf("parse duration %q: %w", s, err)
-	}
-	return d, nil
-}
-
-// startConnOAuthRefresher kicks off the keepalive loop. Called after
-// the toolkit registry + connection store are wired so the resolver
-// can read connection_instances rows. multi-replica is taken from
-// the platform's session-store mode — database-backed sessions
-// implies multi-replica intent.
-func startConnOAuthRefresher(p *platform.Platform) {
-	if p.ConnOAuthStore() == nil {
-		return
-	}
-	store := p.ConnectionStore()
-	if store == nil {
-		return
-	}
-	kinds := buildOAuthKindHandlers(p)
-	if len(kinds) == 0 {
-		return
-	}
-	resolver := &connOAuthConfigResolver{
-		store:     store,
-		kinds:     kinds,
-		maxLifeFn: readMaxLifetime,
-	}
-	multiReplica := p.Config() != nil && p.Config().Sessions.Store == platform.SessionStoreDatabase
-	p.StartConnOAuthRefresher(resolver, multiReplica)
-}
-
-// buildOAuthKindHandlers assembles the per-kind OAuth adapter registry
-// the admin handler dispatches on. Each registered toolkit kind
-// contributes one handler; missing toolkits produce no entry, and the
-// unified handler returns 400 "unsupported connection kind" for
-// requests targeting an unregistered kind.
-func buildOAuthKindHandlers(p *platform.Platform) admin.OAuthKindHandlers {
-	out := admin.OAuthKindHandlers{}
-	if p.ToolkitRegistry() == nil {
-		return out
-	}
-	for _, tk := range p.ToolkitRegistry().All() {
-		switch v := tk.(type) {
-		case *gatewaykit.Toolkit:
-			if h := gatewaykit.NewOAuthKindHandler(v); h != nil {
-				out[connoauth.KindMCP] = h
-			}
-		case *apigatewaykit.Toolkit:
-			out[connoauth.KindAPI] = apigatewaykit.NewOAuthKindHandler(v)
-		}
-	}
-	return out
-}
-
-// wireEnrichmentEngine builds the gateway enrichment engine when a rule
-// store is available, registers the built-in source adapters (Trino,
-// DataHub) bound to the platform's live toolkits, and attaches the
-// engine to the live gateway toolkit so forwarded calls pick it up.
-func wireEnrichmentEngine(p *platform.Platform) *enrichment.Engine {
-	store := p.EnrichmentStore()
-	if store == nil {
-		return nil
-	}
-	sourceReg := enrichment.NewSourceRegistry()
-	registerEnrichmentSources(p, sourceReg)
-
-	engine := enrichment.NewEngine(store, sourceReg)
-	for _, tk := range p.ToolkitRegistry().All() {
-		gw, ok := tk.(*gatewaykit.Toolkit)
-		if !ok {
-			continue
-		}
-		gw.SetEnrichmentEngine(engine)
-	}
-	return engine
-}
-
-// registerEnrichmentSources binds source adapters to the platform's
-// active toolkits. A toolkit that isn't running results in no
-// registration for that source — rules referencing it will surface a
-// "source not registered" warning at evaluation time.
-func registerEnrichmentSources(p *platform.Platform, reg *enrichment.SourceRegistry) {
-	if exec := buildTrinoQueryFunc(p); exec != nil {
-		reg.Register(sources.NewTrinoSource(exec))
-	}
-	if getEntity, getTerm := buildDataHubFuncs(p); getEntity != nil || getTerm != nil {
-		// DataHub source registers even when only a subset of operations
-		// is wired; missing operations report unsupported on dispatch.
-		reg.Register(sources.NewDataHubSource(getEntity, getTerm))
-	}
-}
-
-// buildTrinoQueryFunc returns a TrinoQueryFunc bound to the live trino
-// toolkit's manager, or nil if no trino toolkit is registered.
-func buildTrinoQueryFunc(p *platform.Platform) sources.TrinoQueryFunc {
-	for _, tk := range p.ToolkitRegistry().All() {
-		trinoTk, ok := tk.(*trinokit.Toolkit)
-		if !ok || trinoTk.Manager() == nil {
-			continue
-		}
-		mgr := trinoTk.Manager()
-		return func(ctx context.Context, connection, sql string) ([]map[string]any, error) {
-			c, err := mgr.Client(connection)
-			if err != nil {
-				return nil, fmt.Errorf("trino manager: %w", err)
-			}
-			res, qerr := c.Query(ctx, sql, trinoclient.DefaultQueryOptions())
-			if qerr != nil {
-				return nil, fmt.Errorf("trino query: %w", qerr)
-			}
-			return res.Rows, nil
-		}
-	}
-	return nil
-}
-
-// buildDataHubFuncs returns get-entity and get-glossary-term closures
-// bound to the live datahub toolkit's client, or nils if no datahub
-// toolkit is registered.
-func buildDataHubFuncs(p *platform.Platform) (sources.DataHubGetEntityFunc, sources.DataHubGetGlossaryTermFunc) {
-	for _, tk := range p.ToolkitRegistry().All() {
-		dhTk, ok := tk.(*datahubkit.Toolkit)
-		if !ok || dhTk.Client() == nil {
-			continue
-		}
-		client := dhTk.Client()
-		getEntity := func(ctx context.Context, urn string) (any, error) {
-			return client.GetEntity(ctx, urn)
-		}
-		getTerm := func(ctx context.Context, urn string) (any, error) {
-			return client.GetGlossaryTerm(ctx, urn)
-		}
-		return getEntity, getTerm
-	}
-	return nil, nil
-}
-
-// mountBrowserAuth registers the OIDC login/callback/logout routes.
-func mountBrowserAuth(mux *http.ServeMux, p *platform.Platform) {
-	if p == nil || p.BrowserSessionFlow() == nil {
-		return
-	}
-	flow := p.BrowserSessionFlow()
-	mux.HandleFunc("/portal/auth/login", flow.LoginHandler)
-	mux.HandleFunc("/portal/auth/callback", flow.CallbackHandler)
-	mux.HandleFunc("/portal/auth/logout", flow.LogoutHandler)
-	log.Println("Browser auth enabled (OIDC login on /portal/auth/login)")
-}
-
-// browserRedirectMiddleware redirects browser requests to the portal.
-// Non-browser requests (MCP clients) pass through to the MCP handler.
-func browserRedirectMiddleware(mcpHandler http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == http.MethodGet && strings.Contains(r.Header.Get("Accept"), "text/html") {
-			http.Redirect(w, r, "/portal/", http.StatusTemporaryRedirect)
-			return
-		}
-		mcpHandler.ServeHTTP(w, r)
-	})
-}
-
-// registerOAuthRoutes registers OAuth endpoints on the given mux.
-// Supports both standard paths (with /oauth prefix) and Claude Desktop
-// compatible paths (without /oauth prefix).
-func registerOAuthRoutes(mux *http.ServeMux, handler http.Handler) {
-	// Standard paths (with /oauth prefix)
-	mux.Handle("/.well-known/oauth-authorization-server", handler)
-	mux.Handle("/oauth/authorize", handler)
-	mux.Handle("/oauth/callback", handler)
-	mux.Handle("/oauth/token", handler)
-	mux.Handle("/oauth/register", handler)
-	// Claude Desktop compatibility paths (without /oauth prefix)
-	mux.Handle("/authorize", handler)
-	mux.Handle("/callback", handler)
-	mux.Handle("/token", handler)
-	mux.Handle("/register", handler)
 }
