+++ internal/httpserver/oauth_refresher_helpers_test.go
@@ -1,4 +1,4 @@
-package main
+package httpserver
 
 import (
 	"context"
