+++ internal/httpserver/portal_search_adapter_test.go
@@ -1,4 +1,4 @@
-package main
+package httpserver
 
 import (
 	"context"
