+++ testdata/allowed_internal_imports.txt
@@ -1,32 +1,34 @@
 cmd/dev-mcp-mock -> internal/logsan
-cmd/mcp-data-platform -> internal/apidocs
+cmd/mcp-data-platform -> internal/httpserver
 cmd/mcp-data-platform -> internal/server
-cmd/mcp-data-platform -> internal/ui
-cmd/mcp-data-platform -> pkg/admin
-cmd/mcp-data-platform -> pkg/audit
-cmd/mcp-data-platform -> pkg/browsersession
-cmd/mcp-data-platform -> pkg/connoauth
-cmd/mcp-data-platform -> pkg/gatewayhttp
-cmd/mcp-data-platform -> pkg/health
-cmd/mcp-data-platform -> pkg/http
-cmd/mcp-data-platform -> pkg/knowledge
-cmd/mcp-data-platform -> pkg/observability/proxy
-cmd/mcp-data-platform -> pkg/persona
-cmd/mcp-data-platform -> pkg/pkcestore
 cmd/mcp-data-platform -> pkg/platform
-cmd/mcp-data-platform -> pkg/portal
-cmd/mcp-data-platform -> pkg/portal/datahubapi
-cmd/mcp-data-platform -> pkg/ratelimit
-cmd/mcp-data-platform -> pkg/registry
-cmd/mcp-data-platform -> pkg/resource
-cmd/mcp-data-platform -> pkg/session
-cmd/mcp-data-platform -> pkg/toolkits/apigateway
-cmd/mcp-data-platform -> pkg/toolkits/datahub
-cmd/mcp-data-platform -> pkg/toolkits/gateway
-cmd/mcp-data-platform -> pkg/toolkits/gateway/enrichment
-cmd/mcp-data-platform -> pkg/toolkits/gateway/sources
-cmd/mcp-data-platform -> pkg/toolkits/trino
 cmd/preview-content-viewer -> internal/contentviewer
+internal/httpserver -> internal/apidocs
+internal/httpserver -> internal/ui
+internal/httpserver -> pkg/admin
+internal/httpserver -> pkg/audit
+internal/httpserver -> pkg/browsersession
+internal/httpserver -> pkg/connoauth
+internal/httpserver -> pkg/gatewayhttp
+internal/httpserver -> pkg/health
+internal/httpserver -> pkg/http
+internal/httpserver -> pkg/knowledge
+internal/httpserver -> pkg/observability/proxy
+internal/httpserver -> pkg/persona
+internal/httpserver -> pkg/pkcestore
+internal/httpserver -> pkg/platform
+internal/httpserver -> pkg/portal
+internal/httpserver -> pkg/portal/datahubapi
+internal/httpserver -> pkg/ratelimit
+internal/httpserver -> pkg/registry
+internal/httpserver -> pkg/resource
+internal/httpserver -> pkg/session
+internal/httpserver -> pkg/toolkits/apigateway
+internal/httpserver -> pkg/toolkits/datahub
+internal/httpserver -> pkg/toolkits/gateway
+internal/httpserver -> pkg/toolkits/gateway/enrichment
+internal/httpserver -> pkg/toolkits/gateway/sources
+internal/httpserver -> pkg/toolkits/trino
 internal/platform/browserauth -> pkg/browsersession
 internal/platform/connauth -> pkg/authevents
 internal/platform/connauth -> pkg/connoauth
